[FE] Refactor speaker RE:START, Set Deadline Ticket, Change Desc to Main Topic
