+++ src/app/components/indexes/TicketCarouselRestart25.jsx
@@ -36,7 +36,7 @@ const data = [
     discountPrice: 370000,
     benefit: ["Main stage access", "Workshop", "Experience zone"],
     startDate: "2025-05-28T00:00:00+07:00",
-    expiredDate: "2025-07-26T23:59:00+07:00",
+    expiredDate: "2025-07-12T23:59:00+07:00",
     isPremium: false,
     isActive: true,
   },
@@ -54,7 +54,7 @@ const data = [
       "Exclusive merch set",
     ],
     startDate: "2025-04-27T11:00:00+07:00",
-    expiredDate: "2025-05-27T23:59:00+07:00",
+    expiredDate: "2025-05-12T23:59:00+07:00",
     isPremium: true,
     isActive: true,
   },
@@ -72,7 +72,7 @@ const data = [
       "Exclusive merch set",
     ],
     startDate: "2025-05-28T00:00:00+07:00",
-    expiredDate: "2025-07-25T23:59:00+07:00",
+    expiredDate: "2025-07-12T23:59:00+07:00",
     isPremium: true,
     isActive: true,
   },
