+++ src/app/globals.css
@@ -8,7 +8,7 @@
   --aspect-desktop-banner: 4/1;
   --aspect-leaderboard-banner: 4.55/1;
   --aspect-card-restart: 7/4;
-  --aspect-frame-speaker-restart: 0.89/1;
+  --aspect-speaker-restart: 0.8766/1;
   --color-primary: #0165fc;
   --color-primary-strong: #0041c2;
   --color-primary-dark: #003399;
